use retries instead of wait_for before upload
sometimes icinga isn't ready to accept the newly uploaded file directly
after deleting the old one.

We used to have a wait_for task in there which waited for 10 seconds
before trying to upload again but this wasn't really reliable as well

This changes it to a retries setting on the uri resource instead.